# Tracer performance optimisations (solution 2)