This resolves #254.

This PR handles tracer performance optimisations and it is one of the two suggested solutions which has:

- custom logic in the `tracer.go` file for the JS tracers `call_tracer*`
- **much faster** performance than solution 1, as it doesn't call JS `step` method using duktape if not really needed

Below you can find some sample measurements for each test case, with the important one being the `bigSlow`:

### Test CallTracer

| Test | not optimised | optimised |
|------|-------------:|-----------:|
| bigSlow | 5.45653184s | 51.675229ms |
| create | 2.575857ms | 188.843µs |
| deepCalls | 69.298277ms | 45.796141ms |
| delegatecall | 13.072359ms | 2.932719ms |
| innerCreateOogOuterThrow | 4.356151ms | 723.901µs |
| innerInstafail | 1.789353ms | 1.621216ms |
| innerThrowOuterRevert | 15.972036ms | 6.822914ms |
| oog | 5.654414ms | 124.374µs |
| revert | 2.175ms | 117.949µs |
| revertReason | 4.936105ms | 258.533µs |
| simple | 5.427691ms | 1.60061ms |
| throw | 1.909941ms | 194.468µs |

### Test CallTracerParity

| Test | not optimised | optimised |
|------|-------------:|-----------:|
| bigSlow | 5.377376866s | 28.476724ms |
| create | 2.487462ms | 118.6µs |
| deepCalls | 80.946559ms | 22.162195ms |
| delegatecall | 14.748616ms | 4.620714ms |
| innerCreateOogOuterThrow | 10.448277ms | 576.617µs |
| innerInstafail | 2.620816ms | 733.23µs |
| innerThrowOuterRevert | 15.346743ms | 2.266268ms |
| oog | 5.349929ms | 130.338µs |
| revertReason | 4.295931ms | 99.084µs |
| revert | 1.33ms | 161.985µs |
| simple | 4.94809ms | 1.584925ms |
| throw | 3.851077ms | 167.473µs |